refactor: postprocessors/register.py -> registry.py
